chore: Adding chrom tests, pre-commit, cov
Added pytest and cov to pre-commit.
Added tests for Chromosome module.
Made Parser class an iterator to use next()

refactor: Made Parser class an iterator

Added __next__() method so parser could be
used with next(p) as well as in a loop.

chore: Add pytest and cov to pre-commit

Note to remove these hooks if the tests are too slow

test: Added Chromosome tests